// src/domain/enums/clerk.ts
export enum ClerkEventTypes {
  USER_CREATED = 'user.created',
  USER_DELETED = 'user.deleted',
  USER_UPDATED = 'user.updated',

  ORGANIZATION_CREATED = 'organization.created',
  ORGANIZATION_DELETED = 'organization.deleted',
  ORGANIZATION_UPDATED = 'organization.updated',

  ORGANIZATION_MEMBERSHIP_CREATED = 'organizationMembership.created',
  ORGANIZATION_MEMBERSHIP_DELETED = 'organizationMembership.deleted',
  ORGANIZATION_MEMBERSHIP_UPDATED = 'organizationMembership.updated',
}